test: use pregenerated trees for e2e batch tests (#1528)
* refactor: remove photon_photon client, add missing methods to indexer

* Rename queue_index to prove_by_index in indexer module

* Refactor test code to use `PhotonIndexer` instead of `PhotonClient`.

* format

* feat: add hash field to CompressedAccount

* refactor: update rpc method calls in test_all_endpoints to use get_rpc()

* cleanup

* remove unused imports

* refactor: forester tests: streamline merkle tree account handling

* format

* add Debug to CompressedTransactionTestInputs and update batched state test

Added `#[derive(Debug)]` to `CompressedTransactionTestInputs` in `system_program.rs`. Updated `batched_state_test` by modifying a test case to set `rollover_fee = 1` and included a TODO note for regenerating the batched state Merkle tree accordingly.

* test(batched_address_test.rs): remove leftover

* refactor: unify dependencies to workspace and improve error handling in `from_base58` helper method

+++ sdk-libs/photon-api/src/models/token_data.rs
@@ -13,7 +13,7 @@ use crate::models;
 #[derive(Clone, Default, Debug, PartialEq, Serialize, Deserialize)]
 pub struct TokenData {
     #[serde(rename = "amount")]
-    pub amount: i32,
+    pub amount: u64,
     /// A Solana public key represented as a base58 string.
     #[serde(rename = "delegate", skip_serializing_if = "Option::is_none")]
     pub delegate: Option<String>,
@@ -31,7 +31,7 @@ pub struct TokenData {
 }
 
 impl TokenData {
-    pub fn new(amount: i32, mint: String, owner: String, state: models::AccountState) -> TokenData {
+    pub fn new(amount: u64, mint: String, owner: String, state: models::AccountState) -> TokenData {
         TokenData {
             amount,
             delegate: None,
